requirements.txt
Django>=3.0.6,<3.1
uWSGI>=2.0.18,<2.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
djangorestframework==3.12.2


